created image block component
