+++ lib/gutenberg/gutenberg-editor-component.js
@@ -6,6 +6,7 @@ import * as dataHelper from '../data-helper';
 import AsyncBaseContainer from '../async-base-container';
 import { ContactFormBlockComponent } from './blocks/contact-form-block-component';
 import { ShortcodeBlockComponent } from './blocks/shortcode-block-component';
+import { ImageBlockComponent } from './blocks/image-block-component';
 
 export default class GutenbergEditorComponent extends AsyncBaseContainer {
 	constructor( driver, url ) {
@@ -76,14 +77,13 @@ export default class GutenbergEditorComponent extends AsyncBaseContainer {
 		const blockID = await this.addBlock( 'Shortcode' );
 
 		const shortcodeBlock = await ShortcodeBlockComponent.Expect( this.driver, blockID );
-
 		return await shortcodeBlock.enterShortcode( shortcode );
 	}
 
 	async insertContactForm( email, subject ) {
 		const blockID = await this.addBlock( 'Form' );
-		const contactFormBlock = await ContactFormBlockComponent.Expect( this.driver, blockID );
 
+		const contactFormBlock = await ContactFormBlockComponent.Expect( this.driver, blockID );
 		await contactFormBlock.insertEmail( email );
 		await contactFormBlock.insertSubject( subject );
 		return await contactFormBlock.submitForm();
@@ -131,19 +131,11 @@ export default class GutenbergEditorComponent extends AsyncBaseContainer {
 		return await element.getAttribute( 'value' );
 	}
 
-	async uploadImage( fileDetails ) {
-		await driverHelper.waitTillPresentAndDisplayed(
-			this.driver,
-			By.css( '.components-form-file-upload ' )
-		);
-		const filePathInput = await this.driver.findElement(
-			By.css( '.components-form-file-upload input[type="file"]' )
-		);
-		await filePathInput.sendKeys( fileDetails.file );
-		return await driverHelper.waitTillNotPresent(
-			this.driver,
-			By.css( '.wp-block-image .components-spinner' )
-		); // Wait for upload spinner to complete
+	async addImage( fileDetails ) {
+		const blockID = await this.addBlock( 'Image' );
+
+		const imageBlock = await ImageBlockComponent.Expect( this.driver, blockID );
+		return await imageBlock.uploadImage( fileDetails );
 	}
 
 	async toggleSidebar( open = true ) {
