+++ specs/wp-calypso-gutenberg-page-editor-spec.js
@@ -59,9 +59,8 @@ describe( `[${ host }] Editor: Pages (${ screenSize })`, function() {
 			await gEditorComponent.removeNUXNotice();
 			await gEditorComponent.enterTitle( pageTitle );
 			await gEditorComponent.enterText( pageQuote );
-			await gEditorComponent.addBlock( 'Image' );
+			await gEditorComponent.addImage( fileDetails );
 
-			await gEditorComponent.uploadImage( fileDetails );
 			await gEditorComponent.openSidebar();
 			const gEditorSidebarComponent = await GutenbergEditorSidebarComponent.Expect( driver );
 			await gEditorSidebarComponent.enterImageAltText( fileDetails );
